Improve hardware crypto acceleration support.
Now dev.rtwn.<#>.hwcrypto tunable can be set to one of the following values:
2 - use h/w crypto support for all keys (fine for one-vap configurations,
but will not work for 2 protected vaps (except some edge cases)).
1 - use h/w crypto support for pairwise keys only - will work for all
one-vap and for most two-vap configurations (except when both have
a peer with the same MAC address).
0 - disable h/w crypto support. NOTE: TCP/UDP/IP Rx checksum validation
is not possible in that case.

By default it set to 1 (pairwise in h/w, group in s/w).